Create ethereum mod to share FromWei/TryFromWei trait definitions
